backend/requirements.txt
flask
flask-cors
python-dotenv
pymongo
flask-pymongo